step2 refactoring
