+++ src/js/utils/helper.js
@@ -1,0 +1,11 @@
+const isEmptyString = (string) => {
+  if (string === null) {
+    return true;
+  }
+  if (string.trim() === '') {
+    return true;
+  }
+  return false;
+};
+
+export default isEmptyString;
